fixed view bug

+++ core/src/com/nickdnepr/citadel/util/Constants.java
@@ -5,6 +5,6 @@ import com.badlogic.gdx.Gdx;
 public class Constants {
 
     public static final float DEFAULT_CAMERA_WIDTH = 20;
-    public static final float CAMERA_WIDTH_TO_HEIGHT_COEFFICIENT = Gdx.graphics.getHeight() / Gdx.graphics.getWidth();
+    public static final float CAMERA_WIDTH_TO_HEIGHT_COEFFICIENT = (float) Gdx.graphics.getHeight() / Gdx.graphics.getWidth();
 
 }
